refactor(edge): isolate downstream response emission and result handling
